Migrate Filmstrip data out of ImageModal for improved readability
